+++ webpack/src/index.ts
@@ -1,5 +1,4 @@
 module.exports = {
-    DeckDeckGoMarkdownPlugin: require('./markdown/deckdeckgo-markdown-plugin'),
-    DeckDeckGoInfoPlugin: require('./info/deckdeckgo-info-plugin'),
-    DeckDeckGoRemoveNotesPlugin: require('./notes/deckdeckgo-remove-notes-plugin')
+  DeckDeckGoInfoPlugin: require("./info/deckdeckgo-info-plugin"),
+  DeckDeckGoRemoveNotesPlugin: require("./notes/deckdeckgo-remove-notes-plugin"),
 };